+++ app/src/main/java/org/oxycblt/auxio/detail/recycler/DetailAdapter.kt
@@ -26,9 +26,9 @@ import androidx.recyclerview.widget.RecyclerView
 import org.oxycblt.auxio.IntegerTable
 import org.oxycblt.auxio.databinding.ItemSortHeaderBinding
 import org.oxycblt.auxio.detail.SortHeader
-import org.oxycblt.auxio.ui.recycler.ActivationAdapter
 import org.oxycblt.auxio.ui.recycler.Header
 import org.oxycblt.auxio.ui.recycler.HeaderViewHolder
+import org.oxycblt.auxio.ui.recycler.IndicatorAdapter
 import org.oxycblt.auxio.ui.recycler.Item
 import org.oxycblt.auxio.ui.recycler.MenuItemListener
 import org.oxycblt.auxio.ui.recycler.SimpleItemCallback
@@ -38,7 +38,9 @@ import org.oxycblt.auxio.util.inflater
 abstract class DetailAdapter<L : DetailAdapter.Listener>(
     private val listener: L,
     diffCallback: DiffUtil.ItemCallback<Item>
-) : ActivationAdapter<RecyclerView.ViewHolder>() {
+) : IndicatorAdapter<RecyclerView.ViewHolder>() {
+    private var isPlaying = false
+
     @Suppress("LeakingThis") override fun getItemCount() = differ.currentList.size
 
     override fun getItemViewType(position: Int) =
@@ -77,7 +79,7 @@ abstract class DetailAdapter<L : DetailAdapter.Listener>(
 
     protected val differ = AsyncListDiffer(this, diffCallback)
 
-    val currentList: List<Item>
+    override val currentList: List<Item>
         get() = differ.currentList
 
     fun submitList(list: List<Item>) {