+++ app/src/main/java/org/oxycblt/auxio/home/list/SongListFragment.kt
@@ -29,7 +29,7 @@ import org.oxycblt.auxio.music.Song
 import org.oxycblt.auxio.settings.Settings
 import org.oxycblt.auxio.ui.DisplayMode
 import org.oxycblt.auxio.ui.Sort
-import org.oxycblt.auxio.ui.recycler.ActivationAdapter
+import org.oxycblt.auxio.ui.recycler.IndicatorAdapter
 import org.oxycblt.auxio.ui.recycler.Item
 import org.oxycblt.auxio.ui.recycler.MenuItemListener
 import org.oxycblt.auxio.ui.recycler.SongViewHolder
@@ -58,7 +58,8 @@ class SongListFragment : HomeListFragment<Song>() {
         }
 
         collectImmediately(homeModel.songs, homeAdapter::replaceList)
-        collectImmediately(playbackModel.song, playbackModel.parent, ::handlePlayback)
+        collectImmediately(
+            playbackModel.song, playbackModel.parent, playbackModel.isPlaying, ::handlePlayback)
     }
 
     override fun getPopup(pos: Int): String? {
@@ -113,19 +114,21 @@ class SongListFragment : HomeListFragment<Song>() {
         }
     }
 
-    private fun handlePlayback(song: Song?, parent: MusicParent?) {
+    private fun handlePlayback(song: Song?, parent: MusicParent?, isPlaying: Boolean) {
         if (parent == null) {
-            homeAdapter.activateSong(song)
+            homeAdapter.updateIndicator(song, isPlaying)
         } else {
             // Ignore playback that is not from all songs
-            homeAdapter.activateSong(null)
+            homeAdapter.updateIndicator(null, isPlaying)
         }
     }
 
     private class SongAdapter(private val listener: MenuItemListener) :
-        ActivationAdapter<SongViewHolder>() {
+        IndicatorAdapter<SongViewHolder>() {
         private val differ = SyncListDiffer(this, SongViewHolder.DIFFER)
-        private var currentSong: Song? = null
+
+        override val currentList: List<Item>
+            get() = differ.currentList
 
         override fun getItemCount() = differ.currentList.size
 
@@ -140,20 +143,8 @@ class SongListFragment : HomeListFragment<Song>() {
             }
         }
 
-        override fun shouldActivateViewHolder(position: Int): Boolean {
-            val item = differ.currentList[position]
-            return item.id == currentSong?.id
-        }
-
         fun replaceList(newList: List<Song>) {
             differ.replaceList(newList)
-        }
-
-        /** Update the [song] that this adapter should indicate playback */
-        fun activateSong(song: Song?) {
-            if (song == currentSong) return
-            activateImpl(differ.currentList, currentSong, song)
-            currentSong = song
         }
     }
 }