+++ app/src/main/java/org/oxycblt/auxio/ui/recycler/RecyclerFramework.kt
@@ -173,42 +173,84 @@ abstract class SimpleItemCallback<T : Item> : DiffUtil.ItemCallback<T>() {
     }
 }
 
-abstract class ActivationAdapter<VH : RecyclerView.ViewHolder> : RecyclerView.Adapter<VH>() {
+// TODO: Base adapter that automates current list stuff for span size lookup
+// TODO: Dialog view holder that automates the dumb sizing hack I have to do
+
+abstract class IndicatorAdapter<VH : RecyclerView.ViewHolder> : RecyclerView.Adapter<VH>() {
+    private var isPlaying = false
+    private var currentItem: Item? = null
+
     override fun onBindViewHolder(holder: VH, position: Int) = throw UnsupportedOperationException()
 
     override fun onBindViewHolder(holder: VH, position: Int, payloads: List<Any>) {
-        holder.itemView.isActivated = shouldActivateViewHolder(position)
+        if (holder is IndicatorViewHolder) {
+            val item = currentList[position]
+            val currentItem = currentItem
+            holder.updateIndicator(
+                currentItem != null &&
+                    item.javaClass == currentItem.javaClass &&
+                    item.id == currentItem.id,
+                isPlaying)
+        }
     }
 
-    protected abstract fun shouldActivateViewHolder(position: Int): Boolean
+    abstract val currentList: List<Item>
 
-    protected inline fun <reified T : Item> activateImpl(
-        currentList: List<Item>,
-        oldItem: T?,
-        newItem: T?
-    ) {
-        if (oldItem != null) {
-            val pos = currentList.indexOfFirst { item -> item.id == oldItem.id && item is T }
+    fun updateIndicator(item: Item?, isPlaying: Boolean) {
+        var updatedItem = false
 
-            if (pos > -1) {
-                notifyItemChanged(pos, PAYLOAD_ACTIVATION_CHANGED)
-            } else {
-                logW("oldItem was not in adapter data")
+        if (currentItem != item) {
+            val oldItem = currentItem
+            currentItem = item
+
+            if (oldItem != null) {
+                val pos =
+                    currentList.indexOfFirst {
+                        it.javaClass == oldItem.javaClass && it.id == oldItem.id
+                    }
+
+                if (pos > -1) {
+                    notifyItemChanged(pos, PAYLOAD_INDICATOR_CHANGED)
+                } else {
+                    logW("oldItem was not in adapter data")
+                }
             }
+
+            if (item != null) {
+                val pos =
+                    currentList.indexOfFirst { it.javaClass == item.javaClass && it.id == item.id }
+
+                if (pos > -1) {
+                    notifyItemChanged(pos, PAYLOAD_INDICATOR_CHANGED)
+                } else {
+                    logW("newItem was not in adapter data")
+                }
+            }
+
+            updatedItem = true
         }
 
-        if (newItem != null) {
-            val pos = currentList.indexOfFirst { item -> item is T && item.id == newItem.id }
+        if (this.isPlaying != isPlaying) {
+            this.isPlaying = isPlaying
 
-            if (pos > -1) {
-                notifyItemChanged(pos, PAYLOAD_ACTIVATION_CHANGED)
-            } else {
-                logW("newItem was not in adapter data")
+            if (!updatedItem && item != null) {
+                val pos =
+                    currentList.indexOfFirst { it.javaClass == item.javaClass && it.id == item.id }
+
+                if (pos > -1) {
+                    notifyItemChanged(pos, PAYLOAD_INDICATOR_CHANGED)
+                } else {
+                    logW("newItem was not in adapter data")
+                }
             }
         }
     }
 
     companion object {
-        val PAYLOAD_ACTIVATION_CHANGED = Any()
+        val PAYLOAD_INDICATOR_CHANGED = Any()
     }
+}
+
+abstract class IndicatorViewHolder(root: View) : RecyclerView.ViewHolder(root) {
+    abstract fun updateIndicator(isActive: Boolean, isPlaying: Boolean)
 }