+++ app/src/main/java/org/oxycblt/auxio/ui/recycler/ViewHolders.kt
@@ -37,7 +37,7 @@ import org.oxycblt.auxio.util.inflater
  * @author OxygenCobalt
  */
 class SongViewHolder private constructor(private val binding: ItemSongBinding) :
-    RecyclerView.ViewHolder(binding.root) {
+    IndicatorViewHolder(binding.root) {
     fun bind(item: Song, listener: MenuItemListener) {
         binding.songAlbumCover.bind(item)
         binding.songName.text = item.resolveName(binding.context)
@@ -48,6 +48,11 @@ class SongViewHolder private constructor(private val binding: ItemSongBinding) :
             true
         }
         binding.root.setOnClickListener { listener.onItemClick(item) }
+    }
+
+    override fun updateIndicator(isActive: Boolean, isPlaying: Boolean) {
+        binding.root.isActivated = isActive
+        binding.songAlbumCover.isPlaying = isPlaying
     }
 
     companion object {
@@ -71,7 +76,7 @@ class SongViewHolder private constructor(private val binding: ItemSongBinding) :
 class AlbumViewHolder
 private constructor(
     private val binding: ItemParentBinding,
-) : RecyclerView.ViewHolder(binding.root) {
+) : IndicatorViewHolder(binding.root) {
 
     fun bind(item: Album, listener: MenuItemListener) {
         binding.parentImage.bind(item)
@@ -83,6 +88,11 @@ private constructor(
             true
         }
         binding.root.setOnClickListener { listener.onItemClick(item) }
+    }
+
+    override fun updateIndicator(isActive: Boolean, isPlaying: Boolean) {
+        binding.root.isActivated = isActive
+        binding.parentImage.isPlaying = isPlaying
     }
 
     companion object {
@@ -105,7 +115,7 @@ private constructor(
  * @author OxygenCobalt
  */
 class ArtistViewHolder private constructor(private val binding: ItemParentBinding) :
-    RecyclerView.ViewHolder(binding.root) {
+    IndicatorViewHolder(binding.root) {
 
     fun bind(item: Artist, listener: MenuItemListener) {
         binding.parentImage.bind(item)
@@ -121,6 +131,11 @@ class ArtistViewHolder private constructor(private val binding: ItemParentBindin
             true
         }
         binding.root.setOnClickListener { listener.onItemClick(item) }
+    }
+
+    override fun updateIndicator(isActive: Boolean, isPlaying: Boolean) {
+        binding.root.isActivated = isActive
+        binding.parentImage.isPlaying = isPlaying
     }
 
     companion object {
@@ -145,7 +160,7 @@ class ArtistViewHolder private constructor(private val binding: ItemParentBindin
 class GenreViewHolder
 private constructor(
     private val binding: ItemParentBinding,
-) : RecyclerView.ViewHolder(binding.root) {
+) : IndicatorViewHolder(binding.root) {
 
     fun bind(item: Genre, listener: MenuItemListener) {
         binding.parentImage.bind(item)
@@ -158,6 +173,11 @@ private constructor(
             true
         }
         binding.root.setOnClickListener { listener.onItemClick(item) }
+    }
+
+    override fun updateIndicator(isActive: Boolean, isPlaying: Boolean) {
+        binding.root.isActivated = isActive
+        binding.parentImage.isPlaying = isPlaying
     }
 
     companion object {